Update DOM.XMLNode Class
DOM.XMLNode class is the type of every node in the parsed XML. Here we call getRootElement() to returned the root element of the parsed XML, and assign it to an instance of this class. At this point you have a parsed document. Simply walk through the parse tree by walking by accessing the nodes. The above code calls walkThrough() to do this.